v1.5 -> putnbr now is long

+++ printf/printf.h
@@ -6,7 +6,7 @@
 /*   By: mrubio <mrubio@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/25 02:43:13 by mrubio            #+#    #+#             */
-/*   Updated: 2020/09/25 18:31:14 by mrubio           ###   ########.fr       */
+/*   Updated: 2020/09/25 18:48:49 by mrubio           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,6 @@ void			*ft_memcpy(void *dest, const void *src, unsigned int n);
 char			*ft_strdup(const char *src);
 void			ft_check_arg(char *s, va_list args);
 void			ft_putchar(char c);
-void			ft_putnbr(int n);
+void			ft_putnbr(long n);
 void			ft_putstr(char *s);
 void			ft_putnbr_abs(int n);
